NEWTAG-611

+++ code/src/java/pcgen/core/display/CharacterDisplay.java
@@ -1534,9 +1534,9 @@ public class CharacterDisplay
 		return substitutionClassFacet.get(id, lvl);
 	}
 
-	public Integer getStat(PCStat stat)
+	public int getStat(PCStat stat)
 	{
-		return statValueFacet.get(id, stat);
+		return statValueFacet.get(id, stat).intValue();
 	}
 
 	public boolean containsRacialSubType(RaceSubType st)
